Added login and bid commands
Added commands to login and bid on listings in discord.
Abstracted classes related to listings into listing_classes.py and created a separate file for account classes (account_classes.py).  Both files located under folder "dataclasses". Which replaces the file in the previous commit under the same name.
