+++ goodwill/__init__.py
@@ -1,8 +1,10 @@
 
 BASEURL = "https://buyerapi.shopgoodwill.com/api/"
 
-from goodwill import dataclasses
-from goodwill.dataclasses import Category, Listing
+import goodwill._encryption
+
+import goodwill.dataclasses
+from goodwill.dataclasses import Category, Listing, SimpleListing
 
 from goodwill import db
