+++ goodwill/dataclasses/__init__.py
@@ -1,0 +1,6 @@
+
+import goodwill.dataclasses.listing_classes
+from goodwill.dataclasses.listing_classes import *
+
+import goodwill.dataclasses.account_classes
+from goodwill.dataclasses.account_classes import *
